requirements.txt
elevenlabs
openai
SpeechRecognition
python-dotenv
langchain